Extract test auth service to dedicated file
